@@ -17,6 +17,46 @@ from jaxatari.games.kangaroo_levels import (
     Kangaroo_Level_3,
 )
 
+
+def get_default_asset_config() -> tuple:
+        # 1. Define the game-specific asset manifest in a clear, declarative way.
+        asset_config = [
+            {'name': 'background', 'type': 'background', 'file': 'background.npy'},
+            {
+                'name': 'ape', 'type': 'group',
+                'files': ['ape_standing.npy', 'ape_climb_left.npy', 'ape_moving.npy', 'throwing_ape.npy', 'ape_climb_right.npy']
+            },
+            {
+                'name': 'kangaroo', 'type': 'group',
+                'files': ['kangaroo.npy', 'kangaroo_dead.npy', 'kangaroo_climb.npy', 'kangaroo_ducking.npy', 'kangaroo_jump.npy', 'kangaroo_boxing.npy', 'kangaroo_walk.npy', 'kangaroo_jump_high.npy']
+            },
+            {
+                'name': 'bell', 'type': 'group',
+                'files': ['bell.npy', 'ringing_bell.npy']
+            },
+            {
+                'name': 'fruit', 'type': 'group',
+                'files': ['strawberry.npy', 'tomato.npy', 'cherry.npy', 'pineapple.npy']
+            },
+            {
+                'name': 'child', 'type': 'group',
+                'files': ['child.npy', 'child_jump.npy']
+            },
+            {'name': 'coconut', 'type': 'single', 'file': 'coconut.npy'},
+            {'name': 'thrown_coconut', 'type': 'single', 'file': 'thrown_coconut.npy'},
+            {'name': 'lives', 'type': 'single', 'file': 'kangaroo_lives.npy'},
+            {
+                'name': 'score_digits', 'type': 'digits',
+                'pattern': 'score_{}.npy'
+            },
+            {
+                'name': 'time_digits', 'type': 'digits',
+                'pattern': 'time_{}.npy'
+            }
+        ]
+        return asset_config
+
+
 class KangarooConstants(NamedTuple):
     RESET: int = 18
     RENDER_SCALE_FACTOR: int = 4
@@ -36,7 +76,6 @@ class KangarooConstants(NamedTuple):
     MONKEY_WIDTH: int = 6
     MONKEY_HEIGHT: int = 15
     MONKEY_COLOR: Tuple[int, int, int] = (227, 159, 89)
-    BACKGROUND_COLOR: Tuple[int, int, int] = (80, 0, 132)
     PLAYER_COLOR: Tuple[int, int, int] = (223, 183, 85)
     ENEMY_COLOR: Tuple[int, int, int] = (227, 151, 89)
     FRUIT_COLOR_STATE_1: Tuple[int, int, int] = (214, 92, 92)
@@ -66,6 +105,9 @@ class KangarooConstants(NamedTuple):
     LEVEL_1: LevelConstants = Kangaroo_Level_1
     LEVEL_2: LevelConstants = Kangaroo_Level_2
     LEVEL_3: LevelConstants = Kangaroo_Level_3
+    # sprites to enable asset overrides
+    ASSET_CONFIG: tuple = get_default_asset_config()
+
 
 # -------- Entity Classes --------
 class Entity(NamedTuple):
@@ -187,12 +229,8 @@ class KangarooInfo(NamedTuple):
 
 
 class JaxKangaroo(JaxEnvironment[KangarooState, KangarooObservation, KangarooInfo, KangarooConstants]):
-    def __init__(self, consts: KangarooConstants = None, reward_funcs: list[callable]=None):
+    def __init__(self, consts: KangarooConstants = None):
         super().__init__(consts)
-        self.frame_stack_size = 4
-        if reward_funcs is not None:
-            reward_funcs = tuple(reward_funcs)
-        self.reward_funcs = reward_funcs
         self.action_set = [
             Action.NOOP,
             Action.FIRE,
@@ -625,6 +663,37 @@ class JaxKangaroo(JaxEnvironment[KangarooState, KangarooObservation, KangarooInf
         return jnp.where(has_platform_below, platform_y, jnp.array(1000))
 
     @partial(jax.jit, static_argnums=(0,), donate_argnums=(1,))
+    def _bell_step(self, state: KangarooState) -> Tuple[chex.Array, chex.Array]:
+        """Handles bell collision detection and timer management.
+        
+        Returns:
+            bell_timer: Updated bell timer value
+            respawn_timer_done: Boolean indicating if respawn timer has completed
+        """
+        bell_collision = self._entities_collide(
+            state.player.x,
+            state.player.y,
+            self.consts.PLAYER_WIDTH,
+            state.player.height,
+            state.level.bell_position[0],
+            state.level.bell_position[1],
+            self.consts.BELL_WIDTH,
+            self.consts.BELL_HEIGHT,
+        )
+        bell_active = ~jnp.any(state.level.fruit_stages == 3)
+
+        RESPAWN_AFTER_TICKS = 40
+
+        counter = state.level.bell_timer
+        counter_start = bell_collision & (counter == 0) & bell_active
+        counter = jnp.where(counter_start, 1, counter)
+        counter = jnp.where(counter > 0, counter + 1, counter)
+        counter = jnp.where(counter == RESPAWN_AFTER_TICKS + 1, 0, counter)
+        respawn_timer_done = counter == RESPAWN_AFTER_TICKS
+
+        return counter, respawn_timer_done
+
+    @partial(jax.jit, static_argnums=(0,), donate_argnums=(1,))
     def _fruits_step(self, state: KangarooState) -> Tuple[chex.Array, chex.Array]:
         fruit_x = state.level.fruit_positions[:, 0]
         fruit_y = state.level.fruit_positions[:, 1]
@@ -652,26 +721,7 @@ class JaxKangaroo(JaxEnvironment[KangarooState, KangarooObservation, KangarooInf
         )
         new_score = jnp.sum(score_additions)
 
-        bell_collision = self._entities_collide(
-            state.player.x,
-            state.player.y,
-            self.consts.PLAYER_WIDTH,
-            state.player.height,
-            state.level.bell_position[0],
-            state.level.bell_position[1],
-            self.consts.BELL_WIDTH,
-            self.consts.BELL_HEIGHT,
-        )
-        bell_active = ~jnp.any(state.level.fruit_stages == 3)
-
-        RESPAWN_AFTER_TICKS = 40
-
-        counter = state.level.bell_timer
-        counter_start = bell_collision & (counter == 0) & bell_active
-        counter = jnp.where(counter_start, 1, counter)
-        counter = jnp.where(counter > 0, counter + 1, counter)
-        counter = jnp.where(counter == RESPAWN_AFTER_TICKS + 1, 0, counter)
-        respawn_timer_done = counter == RESPAWN_AFTER_TICKS
+        bell_timer, respawn_timer_done = self._bell_step(state)
 
         def get_new_stages(respawn_timer_done, active, stage):
             return jnp.where(
@@ -690,7 +740,7 @@ class JaxKangaroo(JaxEnvironment[KangarooState, KangarooObservation, KangarooInf
             lambda: new_activations,
         )
 
-        return new_score, activations, new_stages, counter
+        return new_score, activations, new_stages, bell_timer
 
     @partial(jax.jit, static_argnums=(0,), donate_argnums=(1,))
     def _child_step(self, state: KangarooState) -> Tuple[chex.Array]:
@@ -1263,6 +1313,73 @@ class JaxKangaroo(JaxEnvironment[KangarooState, KangarooObservation, KangarooInf
             score_addition,
         )
 
+    @partial(jax.jit, static_argnums=(0,))
+    def _update_coco_state(
+        self,
+        old_m_state: chex.Array,
+        new_m_state: chex.Array,
+        old_m_timer: chex.Array,
+        new_m_timer: chex.Array,
+        c_state: chex.Array,
+        c_pos_x: chex.Array,
+    ) -> chex.Array:
+        return jnp.where(
+            (old_m_state != 3) & (new_m_state == 3),
+            1,
+            jnp.where(
+                (c_state == 1) & (old_m_timer == 3) & (new_m_timer == 2),
+                2,
+                jnp.where(c_pos_x <= 15, 0, c_state),
+            ),
+        )
+
+    @partial(jax.jit, static_argnums=(0,))
+    def _update_coco_positions(
+        self,
+        new_c_state: chex.Array,
+        old_c_state: chex.Array,
+        stepc: chex.Array,
+        old_c_pos: chex.Array,
+        new_m_pos: chex.Array,
+        spawn_position: chex.Array,
+    ) -> chex.Array:
+        return jnp.where(
+            new_c_state == 2,
+            jnp.where(
+                stepc % 2 == 0,
+                jnp.array([old_c_pos[0] - 2, old_c_pos[1]]),
+                old_c_pos,
+            ),
+            jnp.where(
+                (new_c_state == 1) & (old_c_state == 0),
+                jnp.array(
+                    [
+                        new_m_pos[0] - 6,
+                        jnp.where(
+                            spawn_position,
+                            new_m_pos[1] - 5,
+                            new_m_pos[1]
+                            + self.consts.MONKEY_HEIGHT
+                            - self.consts.COCONUT_HEIGHT,
+                        ),
+                    ]
+                ),
+                old_c_pos,
+            ),
+        )
+
+    @partial(jax.jit, static_argnums=(0,), donate_argnums=(1,))
+    def _monkey_controller2(self, state: KangarooState, punching: chex.Array):
+        return (
+            state.level.monkey_states,       # new_monkey_states (all zeros)
+            state.level.monkey_positions,    # new_monkey_positions (all spawn coords/off-screen)
+            state.level.monkey_throw_timers, # new_monkey_throw_timers (all zeros)
+            jnp.zeros((), dtype=jnp.int32),  # score_addition (0)
+            state.level.coco_positions,      # new_coco_positions (all off-screen)
+            state.level.coco_states,         # new_coco_states (all zeros)
+            jnp.array(False),                # flip (should be False)
+        )
+
     @partial(jax.jit, static_argnums=(0,), donate_argnums=(1,))
     def _monkey_controller(self, state: KangarooState, punching: chex.Array):
         current_monkeys_existing = jnp.sum(state.level.monkey_states != 0)
@@ -1316,7 +1433,9 @@ class JaxKangaroo(JaxEnvironment[KangarooState, KangarooObservation, KangarooInf
         )
 
         in_state_1 = new_monkey_states == 1
-        should_transition = (state.level.monkey_positions[:, 1] + self.consts.MONKEY_HEIGHT) >= 172
+        should_transition = (
+            state.level.monkey_positions[:, 1] + self.consts.MONKEY_HEIGHT
+        ) >= 172
         new_monkey_states = jnp.where(
             in_state_1 & should_transition,
             5,
@@ -1437,20 +1556,8 @@ class JaxKangaroo(JaxEnvironment[KangarooState, KangarooObservation, KangarooInf
             state.level.step_counter,
         )
 
-        def update_coco_state(
-            old_m_state, new_m_state, old_m_timer, new_m_timer, c_state, c_pos_x
-        ):
-            return jnp.where(
-                (old_m_state != 3) & (new_m_state == 3),
-                1,
-                jnp.where(
-                    (c_state == 1) & (old_m_timer == 3) & (new_m_timer == 2),
-                    2,
-                    jnp.where(c_pos_x <= 15, 0, c_state),
-                ),
-            )
-
-        new_coco_states = jax.vmap(update_coco_state, in_axes=(0, 0, 0, 0, 0, 0))(
+        # Call the extracted _update_coco_state method
+        new_coco_states = jax.vmap(self._update_coco_state, in_axes=(0, 0, 0, 0, 0, 0))(
             state.level.monkey_states,
             new_monkey_states,
             state.level.monkey_throw_timers,
@@ -1459,39 +1566,19 @@ class JaxKangaroo(JaxEnvironment[KangarooState, KangarooObservation, KangarooInf
             state.level.coco_positions[:, 0],
         )
 
-        def update_coco_positions(new_c_state, old_c_state, stepc, old_c_pos, new_m_pos):
-            return jnp.where(
-                new_c_state == 2,
-                jnp.where(
-                    stepc % 2 == 0,
-                    jnp.array([old_c_pos[0] - 2, old_c_pos[1]]),
-                    old_c_pos,
-                ),
-                jnp.where(
-                    (new_c_state == 1) & (old_c_state == 0),
-                    jnp.array(
-                        [
-                            new_m_pos[0] - 6,
-                            jnp.where(
-                                state.level.spawn_position,
-                                new_m_pos[1] - 5,
-                                new_m_pos[1] + self.consts.MONKEY_HEIGHT - self.consts.COCONUT_HEIGHT,
-                            ),
-                        ]
-                    ),
-                    old_c_pos,
-                ),
-            )
-
-        new_coco_positions = jax.vmap(update_coco_positions, in_axes=(0, 0, None, 0, 0))(
+        # Call the extracted _update_coco_positions method
+        new_coco_positions = jax.vmap(
+            self._update_coco_positions, in_axes=(0, 0, None, 0, 0, None)
+        )(
             new_coco_states,
             state.level.coco_states,
             state.level.step_counter,
             state.level.coco_positions,
             new_monkey_positions,
+            state.level.spawn_position,
         )
 
-        # Handle punching at the very end, after all other state transitions to avoid race conditions
+        # Handle punching at the very end
         fist_w = 3
         fist_h = 4
         fist_x = jnp.where(
@@ -1530,7 +1617,7 @@ class JaxKangaroo(JaxEnvironment[KangarooState, KangarooObservation, KangarooInf
         )
         new_monkey_positions = jax.vmap(
             lambda pos, punched: jnp.where(punched, jnp.array([152, 5]), pos),
-            in_axes=(0, 0)
+            in_axes=(0, 0),
         )(new_monkey_positions, monkeys_punched)
 
         flip = jnp.any((state.level.monkey_states != 3) & (new_monkey_states == 3))
@@ -1976,7 +2063,7 @@ class JaxKangaroo(JaxEnvironment[KangarooState, KangarooObservation, KangarooInf
 
 
 class KangarooRenderer(JAXGameRenderer):
-    def __init__(self, consts=None):
+    def __init__(self, consts: KangarooConstants = None):
         """
         Initializes the renderer by loading sprites, including level backgrounds.
 
@@ -2009,84 +2096,19 @@ class KangarooRenderer(JAXGameRenderer):
         self.ladder_rung_height = 4
         self.ladder_space_height = 4
             
-    
     def _load_sprites(self):
         """Defines the asset manifest for Kangaroo and loads them via the utility function."""
         sprite_path = f"{os.path.dirname(os.path.abspath(__file__))}/sprites/kangaroo"
 
-        # 1. Define the game-specific asset manifest in a clear, declarative way.
-        asset_config = [
-            {'name': 'background', 'type': 'background', 'file': 'background.npy'},
-            {
-                'name': 'ape', 'type': 'group',
-                'files': ['ape_standing.npy', 'ape_climb_left.npy', 'ape_moving.npy', 'throwing_ape.npy', 'ape_climb_right.npy']
-            },
-            {
-                'name': 'kangaroo', 'type': 'group',
-                'files': ['kangaroo.npy', 'kangaroo_dead.npy', 'kangaroo_climb.npy', 'kangaroo_ducking.npy', 'kangaroo_jump.npy', 'kangaroo_boxing.npy', 'kangaroo_walk.npy', 'kangaroo_jump_high.npy']
-            },
-            {
-                'name': 'bell', 'type': 'group',
-                'files': ['bell.npy', 'ringing_bell.npy']
-            },
-            {
-                'name': 'fruit', 'type': 'group',
-                'files': ['strawberry.npy', 'tomato.npy', 'cherry.npy', 'pineapple.npy']
-            },
-            {
-                'name': 'child', 'type': 'group',
-                'files': ['child.npy', 'child_jump.npy']
-            },
-            {'name': 'coconut', 'type': 'single', 'file': 'coconut.npy'},
-            {'name': 'thrown_coconut', 'type': 'single', 'file': 'thrown_coconut.npy'},
-            {'name': 'lives', 'type': 'single', 'file': 'kangaroo_lives.npy'},
-            {
-                'name': 'score_digits', 'type': 'digits',
-                'pattern': 'score_{}.npy'
-            },
-            {
-                'name': 'time_digits', 'type': 'digits',
-                'pattern': 'time_{}.npy'
-            }
-        ]
-
         # 2. Make one call to the utility function. Done.
-        return self.jr.load_and_setup_assets(asset_config, sprite_path)
-        
+        return self.jr.load_and_setup_assets(self.consts.ASSET_CONFIG, sprite_path)
+    
     @partial(jax.jit, static_argnums=(0,))
-    def render(self, state: KangarooState) -> chex.Array:
-        # --- 1. Initialize Raster ---
-        raster = self.jr.create_object_raster(self.BACKGROUND)
-        
-        raster = self.jr.draw_rects(
-            raster,
-            state.level.platform_positions,
-            state.level.platform_sizes,
-            self.PLATFORM_COLOR_ID
-        )
+    def _render_hook_post_ui(self, raster: jnp.ndarray, state: KangarooState):
+        return raster
 
-        raster = self.jr.draw_ladders(
-            raster,
-            state.level.ladder_positions,
-            state.level.ladder_sizes,
-            self.ladder_rung_height,
-            self.ladder_space_height,
-            self.LADDER_COLOR_ID
-        )
-        
-        # --- 3. Draw Dynamic Objects ---
-        # Fruits
-        def _draw_fruit(i, current_raster):
-            should_draw = state.level.fruit_actives[i]
-            fruit_type = state.level.fruit_stages[i].astype(int)
-            pos = state.level.fruit_positions[i]
-            fruit_mask = self.SHAPE_MASKS["fruit"][fruit_type]
-            fruit_offset = self.FLIP_OFFSETS["fruit"]         
-            draw_fn = lambda r: self.jr.render_at(r, pos[0].astype(int), pos[1].astype(int), fruit_mask, flip_offset=fruit_offset)
-            return jax.lax.cond(should_draw, draw_fn, lambda r: r, current_raster)
-        raster = jax.lax.fori_loop(0, state.level.fruit_positions.shape[0], _draw_fruit, raster)
-
-        # Bell
+    @partial(jax.jit, static_argnums=(0,))
+    def _draw_bell(self, raster: jnp.ndarray, state: KangarooState):
         bell_anim_on = ((state.level.bell_animation >= 176) & (state.level.bell_animation <= 192)) | \
                        ((state.level.bell_animation >= 128) & (state.level.bell_animation <= 143)) | \
                        ((state.level.bell_animation >= 80) & (state.level.bell_animation <= 95)) | \
@@ -2100,22 +2122,110 @@ class KangarooRenderer(JAXGameRenderer):
         raster = jax.lax.cond(should_draw_bell,
             lambda r: self.jr.render_at(r, state.level.bell_position[0].astype(int), state.level.bell_position[1].astype(int), bell_mask, flip_horizontal=flip_bell, flip_offset=bell_offset),
             lambda r: r, raster)
+        return raster
+    
+    @partial(jax.jit, static_argnums=(0,))
+    def _draw_ladders(self, raster: jnp.ndarray, state: KangarooState):
+        """Draws the ladders using the utility function."""
+        return self.jr.draw_ladders(
+            raster,
+            state.level.ladder_positions,
+            state.level.ladder_sizes,
+            self.ladder_rung_height,
+            self.ladder_space_height,
+            self.LADDER_COLOR_ID
+        )     
 
-        # Monkeys (Apes)
-        def _draw_monkey(i, current_raster):
+    @partial(jax.jit, static_argnums=(0,))
+    def _draw_single_fruit(self, i, raster, state: KangarooState):
+        """
+        Draws a single fruit based on index i. 
+        Designed to be called within a jax.lax.fori_loop.
+        """
+        should_draw = state.level.fruit_actives[i]
+        fruit_type = state.level.fruit_stages[i].astype(int)
+        pos = state.level.fruit_positions[i]
+        
+        fruit_mask = self.SHAPE_MASKS["fruit"][fruit_type]
+        fruit_offset = self.FLIP_OFFSETS["fruit"]           
+        
+        draw_fn = lambda r: self.jr.render_at(
+            r, 
+            pos[0].astype(int), 
+            pos[1].astype(int), 
+            fruit_mask, 
+            flip_offset=fruit_offset
+        )
+        
+        return jax.lax.cond(should_draw, draw_fn, lambda r: r, raster)
+
+    @partial(jax.jit, static_argnums=(0,))
+    def _draw_single_monkey(self, i, raster, state: KangarooState):
+            """
+            Draws a single monkey based on index i. 
+            Designed to be called within a jax.lax.fori_loop.
+            """
             state_idx = state.level.monkey_states[i].astype(int)
             pos = state.level.monkey_positions[i]
-            monkey_sprite_idx = jnp.array([0, 1, 2, 3, 2, 4])[state_idx] # Map game state to sprite index
+            
+            # Map game state to sprite index
+            monkey_sprite_idx = jnp.array([0, 1, 2, 3, 2, 4])[state_idx] 
+            
             is_walking = (state_idx == 2) | (state_idx == 4)
             use_standing_anim = is_walking & ((state.level.step_counter % 32) < 16)
-            final_sprite_idx = jax.lax.select(use_standing_anim, 0, monkey_sprite_idx) # Index 0 is 'standing'
+            
+            # Index 0 is 'standing'
+            final_sprite_idx = jax.lax.select(use_standing_anim, 0, monkey_sprite_idx) 
+            
             monkey_mask = self.SHAPE_MASKS["ape"][final_sprite_idx]
             flip_offset = self.FLIP_OFFSETS["ape"]
             flip_h = (state_idx == 4)
             should_draw = (state_idx != 0)
-            draw_fn = lambda r: self.jr.render_at_clipped(r, pos[0].astype(int), pos[1].astype(int), monkey_mask, flip_horizontal=flip_h, flip_offset=flip_offset)
-            return jax.lax.cond(should_draw, draw_fn, lambda r: r, current_raster)
-        raster = jax.lax.fori_loop(0, state.level.monkey_positions.shape[0], _draw_monkey, raster)
+            
+            draw_fn = lambda r: self.jr.render_at_clipped(
+                r, 
+                pos[0].astype(int), 
+                pos[1].astype(int), 
+                monkey_mask, 
+                flip_horizontal=flip_h, 
+                flip_offset=flip_offset
+            )
+            
+            return jax.lax.cond(should_draw, draw_fn, lambda r: r, raster)
+
+    @partial(jax.jit, static_argnums=(0,))
+    def render(self, state: KangarooState) -> chex.Array:
+        # --- 1. Initialize Raster ---
+        raster = self.jr.create_object_raster(self.BACKGROUND)
+        
+        raster = self.jr.draw_rects(
+            raster,
+            state.level.platform_positions,
+            state.level.platform_sizes,
+            self.PLATFORM_COLOR_ID
+        )
+
+        raster = self._draw_ladders(raster, state)
+        
+        # --- 3. Draw Dynamic Objects ---
+        # Fruits - UPDATED CALL
+        raster = jax.lax.fori_loop(
+            0, 
+            state.level.fruit_positions.shape[0], 
+            lambda i, r: self._draw_single_fruit(i, r, state), 
+            raster
+        )
+
+        # Bell
+        raster = self._draw_bell(raster, state)
+
+        # Monkeys (Apes)
+        raster = jax.lax.fori_loop(
+            0, 
+            state.level.monkey_positions.shape[0], 
+            lambda i, r: self._draw_single_monkey(i, r, state), 
+            raster
+        )
 
         # Player (Kangaroo)
         is_walking_anim = (state.player.walk_animation > 6) & (state.player.walk_animation < 16) & \
@@ -2182,6 +2292,9 @@ class KangarooRenderer(JAXGameRenderer):
         # Timer
         timer_digits = self.jr.int_to_digits(jnp.maximum(state.level.timer.astype(int), 0), max_digits=4)
         raster = self.jr.render_label(raster, 80, 190, timer_digits, self.SHAPE_MASKS["time_digits"], spacing=4, max_digits=4)
+
+        # Hook for modifications
+        raster = self._render_hook_post_ui(raster, state)
 
         # --- 5. Final Palette Lookup ---
         return self.jr.render_from_palette(raster, self.PALETTE)
