@@ -2,6 +2,7 @@ from jax._src.pjit import JitWrapped
 import os
 from functools import partial
 from typing import NamedTuple, Tuple
+import jax
 import jax.lax
 import jax.numpy as jnp
 import chex
@@ -10,6 +11,26 @@ import jaxatari.spaces as spaces
 from jaxatari.renderers import JAXGameRenderer
 from jaxatari.rendering import jax_rendering_utils as render_utils
 from jaxatari.environment import JaxEnvironment, JAXAtariAction as Action
+
+def _create_wall_sprite(consts: "PongConstants", height: int) -> jnp.ndarray:
+    wall_color_rgba = (*consts.SCORE_COLOR, 255)
+    wall_shape = (height, consts.WIDTH, 4)
+    return jnp.tile(jnp.array(wall_color_rgba, dtype=jnp.uint8), (*wall_shape[:2], 1))
+
+def _get_default_asset_config() -> tuple:
+    """
+    Returns the default declarative asset manifest for Pong.
+    Kept immutable (tuple of dicts) to fit NamedTuple defaults.
+    """
+    return (
+        {'name': 'background', 'type': 'background', 'file': 'background.npy'},
+        {'name': 'player', 'type': 'single', 'file': 'player.npy'},
+        {'name': 'enemy', 'type': 'single', 'file': 'enemy.npy'},
+        {'name': 'ball', 'type': 'single', 'file': 'ball.npy'},
+        {'name': 'player_digits', 'type': 'digits', 'pattern': 'player_score_{}.npy'},
+        {'name': 'enemy_digits', 'type': 'digits', 'pattern': 'enemy_score_{}.npy'},
+    )
+
 
 class PongConstants(NamedTuple):
     MAX_SPEED: int = 12
@@ -38,6 +59,8 @@ class PongConstants(NamedTuple):
     WALL_TOP_HEIGHT: int = 10
     WALL_BOTTOM_Y: int = 194
     WALL_BOTTOM_HEIGHT: int = 16
+    # sset config baked into constants (immutable default) for asset overrides
+    ASSET_CONFIG: tuple = _get_default_asset_config()
 
 
 # immutable state container
@@ -55,6 +78,7 @@ class PongState(NamedTuple):
     step_counter: chex.Array
     acceleration_counter: chex.Array
     buffer: chex.Array
+    key: chex.PRNGKey
 
 
 class EntityPosition(NamedTuple):
@@ -77,13 +101,10 @@ class PongInfo(NamedTuple):
 
 
 class JaxPong(JaxEnvironment[PongState, PongObservation, PongInfo, PongConstants]):
-    def __init__(self, consts: PongConstants = None, reward_funcs: list[callable]=None):
+    def __init__(self, consts: PongConstants = None):
         consts = consts or PongConstants()
         super().__init__(consts)
         self.renderer = PongRenderer(self.consts)
-        if reward_funcs is not None:
-            reward_funcs = tuple(reward_funcs)
-        self.reward_funcs = reward_funcs
         self.action_set = [
             Action.NOOP,
             Action.FIRE,
@@ -92,7 +113,6 @@ class JaxPong(JaxEnvironment[PongState, PongObservation, PongInfo, PongConstants
             Action.RIGHTFIRE,
             Action.LEFTFIRE,
         ]
-        self.obs_size = 3*4+1+1
 
     def _player_step(self, state: PongState, action: chex.Array) -> PongState:
         up = jnp.logical_or(action == Action.LEFT, action == Action.LEFTFIRE)
@@ -195,6 +215,7 @@ class JaxPong(JaxEnvironment[PongState, PongObservation, PongInfo, PongConstants
             step_counter=state.step_counter,
             acceleration_counter=new_acc_counter,
             buffer=buffer,
+            key=state.key,
         )
 
     def _ball_step(self, state: PongState, action) -> PongState:
@@ -326,6 +347,7 @@ class JaxPong(JaxEnvironment[PongState, PongObservation, PongInfo, PongConstants
             step_counter=state.step_counter,
             acceleration_counter=state.acceleration_counter,
             buffer=state.buffer,
+            key=state.key,
         )
 
     def _enemy_step(self, state: PongState) -> PongState:
@@ -351,6 +373,7 @@ class JaxPong(JaxEnvironment[PongState, PongObservation, PongInfo, PongConstants
             step_counter=state.step_counter,
             acceleration_counter=state.acceleration_counter,
             buffer=state.buffer,
+            key=state.key,
         )
 
     def _score_and_reset(self, state: PongState) -> PongState:
@@ -425,7 +448,19 @@ class JaxPong(JaxEnvironment[PongState, PongObservation, PongInfo, PongConstants
             step_counter=step_counter,
             acceleration_counter=state.acceleration_counter,
             buffer=state.buffer,
+            key=state.key,
         )
+        initial_obs = self._get_observation(state)
+
+        return initial_obs, state
+
+    @partial(jax.jit, static_argnums=(0,))
+    def step(self, state: PongState, action: chex.Array) -> Tuple[PongObservation, PongState, float, bool, PongInfo]:
+        previous_state = state
+        state = self._player_step(state, action)
+        state = self._enemy_step(state)
+        state = self._ball_step(state, action)
+        state = self._score_and_reset(state)
 
     def _reset_ball_after_goal(self, state_and_goal: Tuple[PongState, bool]) -> Tuple[chex.Array, chex.Array, chex.Array, chex.Array]:
         state, scored_right = state_and_goal
@@ -447,12 +482,14 @@ class JaxPong(JaxEnvironment[PongState, PongObservation, PongInfo, PongConstants
             ball_vel_y.astype(jnp.int32),
         )
 
-    def reset(self, key=None) -> Tuple[PongObservation, PongState]:
+    def reset(self, key: chex.PRNGKey = jax.random.PRNGKey(42)) -> Tuple[PongObservation, PongState]:
+        # Split key for env reset if needed and for state storage
+        state_key, _step_key = jax.random.split(key)
         state = PongState(
             player_y=jnp.array(96).astype(jnp.int32),
             player_speed=jnp.array(0.0).astype(jnp.int32),
-            ball_x=jnp.array(78).astype(jnp.int32),
-            ball_y=jnp.array(115).astype(jnp.int32),
+            ball_x=self.consts.BALL_START_X.astype(jnp.int32),
+            ball_y=self.consts.BALL_START_Y.astype(jnp.int32),
             enemy_y=jnp.array(115).astype(jnp.int32),
             enemy_speed=jnp.array(0.0).astype(jnp.int32),
             ball_vel_x=self.consts.BALL_SPEED[0].astype(jnp.int32),
@@ -462,6 +499,7 @@ class JaxPong(JaxEnvironment[PongState, PongObservation, PongInfo, PongConstants
             step_counter=jnp.array(0).astype(jnp.int32),
             acceleration_counter=jnp.array(0).astype(jnp.int32),
             buffer=jnp.array(96).astype(jnp.int32),
+            key=state_key,
         )
         initial_obs = self._get_observation(state)
 
@@ -469,11 +507,33 @@ class JaxPong(JaxEnvironment[PongState, PongObservation, PongInfo, PongConstants
 
     @partial(jax.jit, static_argnums=(0,))
     def step(self, state: PongState, action: chex.Array) -> Tuple[PongObservation, PongState, float, bool, PongInfo]:
+        # Split step key from state and keep a new key for the next state
+        new_state_key, step_key = jax.random.split(state.key)
         previous_state = state
+        # Make per-step key available to helpers that may read state.key
+        state = PongState(
+            player_y=state.player_y,
+            player_speed=state.player_speed,
+            ball_x=state.ball_x,
+            ball_y=state.ball_y,
+            enemy_y=state.enemy_y,
+            enemy_speed=state.enemy_speed,
+            ball_vel_x=state.ball_vel_x,
+            ball_vel_y=state.ball_vel_y,
+            player_score=state.player_score,
+            enemy_score=state.enemy_score,
+            step_counter=state.step_counter,
+            acceleration_counter=state.acceleration_counter,
+            buffer=state.buffer,
+            key=step_key,
+        )
         state = self._player_step(state, action)
         state = self._enemy_step(state)
         state = self._ball_step(state, action)
         state = self._score_and_reset(state)
+
+        # Update state key to new_state_key for next step
+        state = state._replace(key=new_state_key)
 
         done = self._get_done(state)
         env_reward = self._get_reward(previous_state, state)
@@ -589,7 +649,7 @@ class JaxPong(JaxEnvironment[PongState, PongObservation, PongInfo, PongConstants
 
 class PongRenderer(JAXGameRenderer):
     def __init__(self, consts: PongConstants = None):
-        super().__init__()
+        super().__init__(consts)
         self.consts = consts or PongConstants()
         self.config = render_utils.RendererConfig(
             game_dimensions=(210, 160),
@@ -597,43 +657,27 @@ class PongRenderer(JAXGameRenderer):
             #downscale=(84, 84)
         )
         self.jr = render_utils.JaxRenderingUtils(self.config)
-        # 1. Create procedural assets for both walls
-        wall_sprite_top = self._create_wall_sprite(self.consts.WALL_TOP_HEIGHT)
-        wall_sprite_bottom = self._create_wall_sprite(self.consts.WALL_BOTTOM_HEIGHT)
-        
-        # 2. Update asset config to include both walls
-        asset_config = self._get_asset_config(wall_sprite_top, wall_sprite_bottom)
-        sprite_path = f"{os.path.dirname(os.path.abspath(__file__))}/sprites/pong"
 
-        # 3. Make a single call to the setup function
+        # 1. Start from (possibly modded) asset config provided via constants
+        final_asset_config = list(self.consts.ASSET_CONFIG)
+
+        # 2. Create procedural assets using modded constants
+        wall_sprite_top = _create_wall_sprite(self.consts, self.consts.WALL_TOP_HEIGHT)
+        wall_sprite_bottom = _create_wall_sprite(self.consts, self.consts.WALL_BOTTOM_HEIGHT)
+
+        # 3. Append procedural assets
+        final_asset_config.append({'name': 'wall_top', 'type': 'procedural', 'data': wall_sprite_top})
+        final_asset_config.append({'name': 'wall_bottom', 'type': 'procedural', 'data': wall_sprite_bottom})
+
+        # 4. Bake assets once
+        sprite_path = f"{os.path.dirname(os.path.abspath(__file__))}/sprites/pong"
         (
             self.PALETTE,
             self.SHAPE_MASKS,
             self.BACKGROUND,
             self.COLOR_TO_ID,
             self.FLIP_OFFSETS
-        ) = self.jr.load_and_setup_assets(asset_config, sprite_path)
-
-    def _create_wall_sprite(self, height: int) -> jnp.ndarray:
-        """Procedurally creates an RGBA sprite for a wall of given height."""
-        wall_color_rgba = (*self.consts.SCORE_COLOR, 255) # e.g., (236, 236, 236, 255)
-        wall_shape = (height, self.consts.WIDTH, 4)
-        wall_sprite = jnp.tile(jnp.array(wall_color_rgba, dtype=jnp.uint8), (*wall_shape[:2], 1))
-        return wall_sprite
-
-    def _get_asset_config(self, wall_sprite_top: jnp.ndarray, wall_sprite_bottom: jnp.ndarray) -> list:
-        """Returns the declarative manifest of all assets for the game, including both wall sprites."""
-        return [
-            {'name': 'background', 'type': 'background', 'file': 'background.npy'},
-            {'name': 'player', 'type': 'single', 'file': 'player.npy'},
-            {'name': 'enemy', 'type': 'single', 'file': 'enemy.npy'},
-            {'name': 'ball', 'type': 'single', 'file': 'ball.npy'},
-            {'name': 'player_digits', 'type': 'digits', 'pattern': 'player_score_{}.npy'},
-            {'name': 'enemy_digits', 'type': 'digits', 'pattern': 'enemy_score_{}.npy'},
-            # Add the procedurally created sprites to the manifest
-            {'name': 'wall_top', 'type': 'procedural', 'data': wall_sprite_top},
-            {'name': 'wall_bottom', 'type': 'procedural', 'data': wall_sprite_bottom},
-        ]
+        ) = self.jr.load_and_setup_assets(final_asset_config, sprite_path)
 
     @partial(jax.jit, static_argnums=(0,))
     def render(self, state):
