@@ -10,6 +10,13 @@ from jaxatari.renderers import JAXGameRenderer
 from jaxatari.rendering import jax_rendering_utils as render_utils
 import jaxatari.spaces as spaces
 
+def _get_default_asset_config() -> tuple:
+    """
+    Returns the default declarative asset manifest for Surround.
+    Kept immutable (tuple of dicts) to fit NamedTuple defaults.
+    Note: Surround uses mostly procedural sprites, so this is empty.
+    """
+    return ()
 
 class SurroundConstants(NamedTuple):
     """Parameters defining the Surround grid and visuals."""
@@ -65,6 +72,9 @@ class SurroundConstants(NamedTuple):
     # Rough logic rate control when caller steps at ~60 FPS
     # Move only every N calls to step (e.g., 60/4 = 15 for ~4 Hz)
     MOVE_EVERY_N_STEPS: int = 15
+
+    # Asset config baked into constants (immutable default) for asset overrides
+    ASSET_CONFIG: tuple = _get_default_asset_config()
     # --- Speed-up schedule (game accelerates over time) ---
     # Every SPEEDUP_STEPS logic ticks, reduce the effective MOVE_EVERY by SPEEDUP_DELTA, but not below MIN_MOVE_EVERY.
     # SPEEDUP_STEPS: int = 200          # how many logic ticks until next speed bump
@@ -123,7 +133,6 @@ class JaxSurround(
     def __init__(
         self,
         consts: Optional[SurroundConstants] = None,
-        reward_funcs: Optional[Sequence[Callable[[SurroundState, SurroundState], jnp.ndarray]]] = None,
     ):
         consts = consts or SurroundConstants()
         super().__init__(consts)
@@ -136,8 +145,6 @@ class JaxSurround(
             Action.LEFT,
             Action.DOWN,
         ]
-        # Bleibt während der Laufzeit statisch -> JAX-jit-freundlich.
-        self.reward_funcs = reward_funcs
 
     # --- Internal AI helper for P1 (left player) ---
 
@@ -597,33 +604,16 @@ class SurroundRenderer(JAXGameRenderer):
         )
         self.jr = render_utils.JaxRenderingUtils(self.config)
 
-        # --- FIX: Centralize all estimated colors as instance attributes ---
         self.P1_HEAD_COLOR_TUPLE = (214, 214, 42)    # Yellow
         self.P2_HEAD_COLOR_TUPLE = (198, 89, 179)    # Red/Pink
         self.PLAYFIELD_COLOR_TUPLE = (181, 119, 181) # Lavender
         self.BORDER_COLOR_TUPLE = (214, 92, 92)      # Pink
         self.DIVIDER_COLOR_TUPLE = (142, 142, 142)   # Grey
 
-        asset_config = self._get_asset_config()
-        (
-            self.PALETTE,
-            self.SHAPE_MASKS,
-            self.BACKGROUND,
-            self.COLOR_TO_ID,
-            self.FLIP_OFFSETS,
-        ) = self.jr.load_and_setup_assets(asset_config, "")
-
-        self.TRAIL_COLOR_MAP = jnp.array([
-            self.jr.TRANSPARENT_ID,
-            self.COLOR_TO_ID[self.consts.P1_TRAIL_COLOR],
-            self.COLOR_TO_ID[self.consts.P2_TRAIL_COLOR],
-        ])
-
-    def _get_asset_config(self) -> list:
-        """Returns the declarative manifest of all assets for the game."""
-        config = []
-
-        # --- FIX: Use the instance attributes defined in __init__ ---
+        # 1. Start from (possibly modded) asset config provided via constants
+        final_asset_config = list(self.consts.ASSET_CONFIG)
+        
+        # 2. Create procedural assets using modded constants
         procedural_sprites = {
             'p1_head': jnp.array(list(self.P1_HEAD_COLOR_TUPLE) + [255], dtype=jnp.uint8).reshape(1, 1, 4),
             'p2_head': jnp.array(list(self.P2_HEAD_COLOR_TUPLE) + [255], dtype=jnp.uint8).reshape(1, 1, 4),
@@ -634,10 +624,10 @@ class SurroundRenderer(JAXGameRenderer):
             'playfield': jnp.array(list(self.PLAYFIELD_COLOR_TUPLE) + [255], dtype=jnp.uint8).reshape(1, 1, 4),
         }
 
-        config.append({'name': 'background', 'type': 'background', 'data': jnp.array([0, 0, 0, 255], dtype=jnp.uint8).reshape(1, 1, 4)})
+        final_asset_config.append({'name': 'background', 'type': 'background', 'data': jnp.array([0, 0, 0, 255], dtype=jnp.uint8).reshape(1, 1, 4)})
 
         for name, data in procedural_sprites.items():
-            config.append({'name': name, 'type': 'procedural', 'data': data})
+            final_asset_config.append({'name': name, 'type': 'procedural', 'data': data})
 
         module_dir = os.path.dirname(os.path.abspath(__file__))
         digit_path = os.path.join(module_dir, "sprites/seaquest/digits/{}" + ".npy")
@@ -650,10 +640,23 @@ class SurroundRenderer(JAXGameRenderer):
         p1_digits_rgba = jnp.concatenate([jnp.where(alpha_mask, jnp.array(self.P1_HEAD_COLOR_TUPLE), 0), scaled_digits[..., 3:]], axis=-1)
         p2_digits_rgba = jnp.concatenate([jnp.where(alpha_mask, jnp.array(self.P2_HEAD_COLOR_TUPLE), 0), scaled_digits[..., 3:]], axis=-1)
 
-        config.append({'name': 'p1_digits', 'type': 'procedural', 'data': p1_digits_rgba})
-        config.append({'name': 'p2_digits', 'type': 'procedural', 'data': p2_digits_rgba})
+        final_asset_config.append({'name': 'p1_digits', 'type': 'procedural', 'data': p1_digits_rgba})
+        final_asset_config.append({'name': 'p2_digits', 'type': 'procedural', 'data': p2_digits_rgba})
+        
+        # 3. Load all assets, create palette, and generate ID masks
+        (
+            self.PALETTE,
+            self.SHAPE_MASKS,
+            self.BACKGROUND,
+            self.COLOR_TO_ID,
+            self.FLIP_OFFSETS,
+        ) = self.jr.load_and_setup_assets(final_asset_config, "")
 
-        return config
+        self.TRAIL_COLOR_MAP = jnp.array([
+            self.jr.TRANSPARENT_ID,
+            self.COLOR_TO_ID[self.consts.P1_TRAIL_COLOR],
+            self.COLOR_TO_ID[self.consts.P2_TRAIL_COLOR],
+        ])
 
     @partial(jax.jit, static_argnums=(0,))
     def render(self, state: SurroundState) -> jnp.ndarray:
@@ -679,7 +682,6 @@ class SurroundRenderer(JAXGameRenderer):
         by = self.consts.BORDER_CELLS_Y * cell_h
         border_positions = jnp.array([[0, y_off], [0, y_off + field_h - by], [0, y_off], [field_w - bx, y_off]])
         border_sizes = jnp.array([[field_w, by], [field_w, by], [bx, field_h], [bx, field_h]])
-        # --- FIX: Use the consistent color attribute for the border ---
         raster = self.jr.draw_rects(raster, border_positions, border_sizes, self.COLOR_TO_ID[self.BORDER_COLOR_TUPLE])
 
         occupied_grid = jnp.logical_or(state.trail != 0, state.border).T.astype(jnp.int32)
@@ -693,7 +695,6 @@ class SurroundRenderer(JAXGameRenderer):
         divider_thickness = max(1, self.consts.DIVIDER_THICKNESS)
         band_mask = (relative_y % cell_h >= mid) & (relative_y % cell_h < mid + divider_thickness)
         final_divider_mask = jnp.logical_and(grid_mask_raster, band_mask)
-        # --- FIX: Use the consistent color attribute for the divider ---
         raster = jnp.where(final_divider_mask, self.COLOR_TO_ID[self.DIVIDER_COLOR_TUPLE], raster)
 
         p1x = state.pos0[0] * cell_w
@@ -703,7 +704,6 @@ class SurroundRenderer(JAXGameRenderer):
 
         p1_trail_mask = jnp.ones((cell_h, cell_w), dtype=jnp.uint8) * self.COLOR_TO_ID[self.consts.P1_TRAIL_COLOR]
         p2_trail_mask = jnp.ones((cell_h, cell_w), dtype=jnp.uint8) * self.COLOR_TO_ID[self.consts.P2_TRAIL_COLOR]
-        # --- FIX: Use the consistent color attributes for the heads ---
         p1_head_mask = jnp.ones((cell_h, cell_w), dtype=jnp.uint8) * self.COLOR_TO_ID[self.P1_HEAD_COLOR_TUPLE]
         p2_head_mask = jnp.ones((cell_h, cell_w), dtype=jnp.uint8) * self.COLOR_TO_ID[self.P2_HEAD_COLOR_TUPLE]
 
