@@ -14,6 +14,21 @@ from jaxatari.rendering import jax_rendering_utils as render_utils
 import numpy as np
 from jaxatari.renderers import JAXGameRenderer
 
+def _get_default_asset_config() -> tuple:
+    """
+    Returns the default declarative asset manifest for Tetris.
+    Kept immutable (tuple of dicts) to fit NamedTuple defaults.
+    """
+    return (
+        {'name': 'background', 'type': 'background', 'file': 'background.npy'},
+        {'name': 'board_overlay', 'type': 'single', 'file': 'board.npy'},
+        {'name': 'score_digits', 'type': 'digits', 'pattern': 'score/score_{}.npy'},
+        {'name': 'banner_one', 'type': 'single', 'file': 'text_one.npy'},
+        {'name': 'banner_two', 'type': 'single', 'file': 'text_two.npy'},
+        {'name': 'banner_triple', 'type': 'single', 'file': 'text_triple.npy'},
+        {'name': 'banner_tetris', 'type': 'single', 'file': 'text_tetris.npy'},
+    )
+
 class TetrisConstants(NamedTuple):
     # logical grid (Board)
     BOARD_WIDTH: int = 10
@@ -91,6 +106,9 @@ class TetrisConstants(NamedTuple):
     ], dtype=jnp.int32)
     RESET: int = Action.DOWNLEFTFIRE  # keep a reserved reset action
 
+    # Asset config baked into constants (immutable default) for asset overrides
+    ASSET_CONFIG: tuple = _get_default_asset_config()
+
 # ======================== State/Obs/Info =================
 class TetrisState(NamedTuple):
 
@@ -134,14 +152,13 @@ class TetrisInfo(NamedTuple):
 # ======================= Environment =====================
 
 class JaxTetris(JaxEnvironment[TetrisState, TetrisObservation, TetrisInfo, TetrisConstants]):
-    def __init__(self, consts: TetrisConstants = None, reward_funcs: list[callable]=None, instant_drop: bool = False):
+    def __init__(self, consts: TetrisConstants = None):
         """ Initialize the JaxTetris environment"""
 
         consts = consts or TetrisConstants()
         super().__init__(consts)
         self.renderer = TetrisRenderer(self.consts)
-        self.instant_drop = instant_drop
-        self.reward_funcs = reward_funcs
+        self.instant_drop = False
 
     # ----- Helpers -----
     @partial(jax.jit, static_argnums=0)
@@ -595,47 +612,35 @@ class TetrisRenderer(JAXGameRenderer):
         )
         self.jr = render_utils.JaxRenderingUtils(self.config)
 
-        # 1. Load all assets using the declarative pattern
-        asset_config = self._get_asset_config()
+        # 1. Start from (possibly modded) asset config provided via constants
+        final_asset_config = list(self.consts.ASSET_CONFIG)
+        
+        # 2. Create procedural assets using modded constants
+        # Procedurally generate 1x1 pixel sprites for each of the 22 row colors.
+        # This ensures they are all included in the final color palette.
         sprite_path = f"{os.path.dirname(os.path.abspath(__file__))}/sprites/tetris"
-
+        for i in range(22):
+            # The file is loaded here to get the color, then converted to a procedural sprite.
+            color_rgba = self.jr.loadFrame(
+                f"{sprite_path}/height_colors/h_{i}.npy"
+            )[0, 0] # Get the color from the top-left pixel
+            final_asset_config.append({
+                'name': f'row_color_{i}',
+                'type': 'procedural',
+                'data': color_rgba.reshape(1, 1, 4)
+            })
+        
+        # 3. Load all assets, create palette, and generate ID masks
         (
             self.PALETTE,
             self.SHAPE_MASKS,
             self.BACKGROUND,
             self.COLOR_TO_ID,
             self.FLIP_OFFSETS,
-        ) = self.jr.load_and_setup_assets(asset_config, sprite_path)
+        ) = self.jr.load_and_setup_assets(final_asset_config, sprite_path)
 
-        # 2. Precompute the color map for the board grid
+        # 4. Precompute the color map for the board grid
         self.BOARD_COLOR_MAP = self._precompute_board_color_map()
-
-    def _get_asset_config(self) -> list:
-        """Returns the declarative manifest of all assets for the game."""
-        # This list defines all sprites that need to be loaded from files.
-        config = [
-            {'name': 'background', 'type': 'background', 'file': 'background.npy'},
-            {'name': 'board_overlay', 'type': 'single', 'file': 'board.npy'},
-            {'name': 'score_digits', 'type': 'digits', 'pattern': 'score/score_{}.npy'},
-            {'name': 'banner_one', 'type': 'single', 'file': 'text_one.npy'},
-            {'name': 'banner_two', 'type': 'single', 'file': 'text_two.npy'},
-            {'name': 'banner_triple', 'type': 'single', 'file': 'text_triple.npy'},
-            {'name': 'banner_tetris', 'type': 'single', 'file': 'text_tetris.npy'},
-        ]
-
-        # Procedurally generate 1x1 pixel sprites for each of the 22 row colors.
-        # This ensures they are all included in the final color palette.
-        for i in range(22):
-            # The file is loaded here to get the color, then converted to a procedural sprite.
-            color_rgba = self.jr.loadFrame(
-                f"{os.path.dirname(os.path.abspath(__file__))}/sprites/tetris/height_colors/h_{i}.npy"
-            )[0, 0] # Get the color from the top-left pixel
-            config.append({
-                'name': f'row_color_{i}',
-                'type': 'procedural',
-                'data': color_rgba.reshape(1, 1, 4)
-            })
-        return config
 
     def _precompute_board_color_map(self) -> jnp.ndarray:
         """Creates a lookup table mapping a row's object ID to its color ID."""
